pyproject.toml
[project]
name = "mental-health-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.10",
    "langchain-core>=1.2.12",
    "langchain-openai>=1.1.9",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
